backend/requirements.txt
fastapi
uvicorn
openai
tavily-python
python-dotenv
requests
pydantic
httpx
